% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_analysis.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calcular Metricas Comportamentais Universais}
\usage{
calculate_metrics(processed_heatmap, block_size = 8)
}
\arguments{
\item{processed_heatmap}{Objeto retornado pela funcao \code{process_heatmap}.}

\item{block_size}{Tamanho do bloco em pixels para a grade do Gini (padrao 8).}
}
\value{
Uma lista com as 4 metricas calculadas e tabelas resumo.
}
\description{
Calcula o Indice de Tigmotaxia, Gini Comportamental, Distancia Media ao Centro
e o Centroide de Permanencia ponderados pela intensidade do mapa de calor.
}
