% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_analysis.R
\name{plot_behavior}
\alias{plot_behavior}
\title{Plotar Analise Espacial Completa}
\usage{
plot_behavior(processed_heatmap, metrics)
}
\arguments{
\item{processed_heatmap}{Objeto retornado pela funcao \code{process_heatmap}.}

\item{metrics}{Objeto retornado pela funcao \code{calculate_metrics}.}
}
\value{
Um grafico ggplot2.
}
\description{
Gera um grafico contendo o rastro do animal, a divisao da arena e o
Centroide de Permanencia (marcado como uma estrela vermelha).
}
