% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_analysis.R
\name{process_heatmap}
\alias{process_heatmap}
\title{Processar Heatmap e Extrair Geometria da Arena}
\usage{
process_heatmap(file_path, proportion_center = 0.8)
}
\arguments{
\item{file_path}{Caminho para o arquivo de imagem do heatmap.}

\item{proportion_center}{Proporcao do raio para definir a zona central (padrao 0.8).}
}
\value{
Uma lista com o dataframe de pixels ativos e metadados geometricos.
}
\description{
Esta funcao le a imagem do mapa de calor (PNG/JPEG), remove o fundo branco
e calcula as coordenadas do centro geometrico e raio util da arena.
}
